refactor(common): regroup physics into common/physics/thermodynamics, tests follow
From OngChia's review: "AES" out of every docstring under common/physics, the
stencils grouped by the quantity they compute, and the host-side numpy helpers
of common/thermodynamic_functions.py brought into the same package.

    thermodynamics/compute_pressure.py     surface + hydrostatic pressure (stencils
                                           and host sequencing), saturation vapour
                                           pressures
    thermodynamics/compute_temperature.py  virtual temperature and temperature; the
                                           exner/theta_v update from a Tv tendency
    thermodynamics/compute_energy.py       internal energy <-> temperature
    thermodynamics/compute_moisture.py     specific humidity, qv from relative humidity
    thermodynamics/compute_tendencies.py   the three tendency stencils

`compute_tendencies` is a fifth module beyond the four proposed: the tendencies
fit none of the quantity buckets and muphys imports them as a unit. A stencil
that leaves a `stencils/` directory loses the per-file PLR0917 exemption and
carries the usual bracketed noqa instead.

Tests follow the modules: one stencil-test file per module, class names after
the programs. The datatests move from unit_tests/ to integration_tests/ as in
dycore and diffusion; the RBF wind datatest goes to
interpolation/integration_tests, where its stencil lives, and the plain unit
test of update_exner_and_theta_v to unit_tests/.

Also: `diagnose_pressure_surface_to_top` -> `compute_surface_and_hydrostatic_pressure`
(and `_ndarray`), the `diagnose_*` import aliases dropped, and three comments in
jablonowski_williamson.py restored to main's wording -- the port's expanded
versions had been taken along with the file.
